Move Mermaid depth ordering helpers to internal types
