+++ src/SlnxMermaid.Core/Emit/DepthOrderedEdge.cs
@@ -1,0 +1,19 @@
+﻿namespace SlnxMermaid.Core.Emit
+{
+internal sealed class DepthOrderedEdge
+{
+    internal DepthOrderedEdge(
+        string rootId,
+        string from,
+        string to)
+    {
+        RootId = rootId;
+        From = from;
+        To = to;
+    }
+
+    internal string RootId { get; }
+    internal string From { get; }
+    internal string To { get; }
+}
+}
