+++ src/SlnxMermaid.Core/Emit/LegacyEdge.cs
@@ -1,0 +1,14 @@
+﻿namespace SlnxMermaid.Core.Emit
+{
+internal sealed class LegacyEdge
+{
+    internal LegacyEdge(string from, string to)
+    {
+        From = from;
+        To = to;
+    }
+
+    internal string From { get; }
+    internal string To { get; }
+}
+}
